float amount

+++ lnurl.py
@@ -115,7 +115,7 @@ async def lnurl_params(
         price_msat = (
             await fiat_amount_as_satoshis(float(amount), device.currency)
             if device.currency != "sat"
-            else int(amount)
+            else float(amount)
         ) * 1000
 
         # Check they're not trying to trick the switch!
